Yarnspinner dialogs trigger on click
Added a Mouse Left Button click interact to the player controller. Added a crosshair. Brought YarnSpinner basics into Prototyping Scene. Hooked player controller up so that it checks to see if it's looking at a GameObject that has YarnInteractable.cs attached to it. If so, it tells YarnInteractable to start the specified node. Also made it so that GameObjects with YarnInteractable on them cause the crosshair to turn green when it's "looking" at those objects.